+++ system/src/Grav/Common/Twig/TokenParser/TwigTokenParserStyle.php
@@ -56,13 +56,29 @@ class TwigTokenParserStyle extends AbstractTokenParser
     {
         $stream = $this->parser->getStream();
 
+        // Look for deprecated {% style ... in ... %}
+        if (!$stream->test(Token::BLOCK_END_TYPE) && !$stream->test(Token::OPERATOR_TYPE, 'in')) {
+            $i = 0;
+            do {
+                $token = $stream->look(++$i);
+                if ($token->test(Token::BLOCK_END_TYPE)) {
+                    break;
+                }
+                if ($token->test(Token::OPERATOR_TYPE, 'in') && $stream->look($i+1)->test(Token::STRING_TYPE)) {
+                    user_error("Twig: Using {% style ... in ... %} is deprecated, use {% style ...  at ... %} instead", E_USER_DEPRECATED);
+
+                    break;
+                }
+            } while (true);
+        }
+
         $file = null;
-        if (!$stream->test(Token::NAME_TYPE) && !$stream->test(Token::OPERATOR_TYPE) && !$stream->test(Token::BLOCK_END_TYPE)) {
+        if (!$stream->test(Token::NAME_TYPE) && !$stream->test(Token::OPERATOR_TYPE, 'in') && !$stream->test(Token::BLOCK_END_TYPE)) {
             $file = $this->parser->getExpressionParser()->parseExpression();
         }
 
         $group = null;
-        if ($stream->nextIf(Token::OPERATOR_TYPE, 'in')) {
+        if ($stream->nextIf(Token::NAME_TYPE, 'at') || $stream->nextIf(Token::OPERATOR_TYPE, 'in')) {
             $group = $this->parser->getExpressionParser()->parseExpression();
         }